重构：admin shared.py 拆分为 shared 包（行为不变）
按职责拆分为 _runtime/_oopz/_area/_debug/_netease/_bilibili/_pages/_session/_snapshots，__init__.py 聚合并复刻原顶层导入与 __all__，保持 from web.admin.shared import * 完全兼容。同时修正 set-cookie 回退分支：新增 _cookie_pairs_from_header 仅提取 name=value、丢弃属性段。
